JARVIS_MASTER/requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
python-telegram-bot>=20.0
rich>=13.7.0
requests>=2.31.0
psutil>=5.9.8
Pillow>=10.3.0
duckduckgo-search>=6.1.0
pyperclip>=1.8.2
colorama>=0.4.6
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
pyautogui>=0.9.54
